Add '--format' and '--output' option
Add some options for the 'geekhours course list' and
'geekhours done list' commands.

The '--format' and '-f' options format data to comma separated CSV or
JSON and write it to a file.

File path can be specified by giving it to the '--output' or '-o'
option. If these options are not specified, the following file names
will be used by default.

For the 'geekhours course list' command, 'courses.csv' or 'courses.json'
will be used, and for the 'geekhours done list' command, 'records.csv'
or 'records.json' will be used depending on the format type.